Move complex fetch parse logic to helper
