+++ client/src/helpers/fetch.js
@@ -1,0 +1,32 @@
+// Copyright 2023 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+
+export const getDjangoError = data =>  {
+    // try and parse out something more from Django's debugging screen
+    let errorText = ""
+
+    if (data.includes('exception_value')) {
+        // css class from Django default debug screen
+        let myDoc = new DOMParser();
+        let djDoc = myDoc.parseFromString(data, 'text/html');
+        let djError =
+            djDoc.getElementsByClassName('exception_value')[0].innerText;
+
+        errorText = `Django Debug: "${djError}"`;
+    }
+    return errorText
+}
+
+export default { getDjangoError };
